+++ internal/embeddings/embeddings_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func TestCosineSimilarityIdentical(t *testing.T) {
-	vector := []float32{1.0, 2.0, 3.0}
+	vector := []float64{1.0, 2.0, 3.0}
 	similarity := CosineSimilarity(vector, vector)
 	if math.Abs(similarity-1.0) > 1e-6 {
 		t.Errorf("identical vectors should have similarity 1.0, got %f", similarity)
@@ -14,8 +14,8 @@ func TestCosineSimilarityIdentical(t *testing.T) {
 }
 
 func TestCosineSimilarityOrthogonal(t *testing.T) {
-	vectorA := []float32{1.0, 0.0}
-	vectorB := []float32{0.0, 1.0}
+	vectorA := []float64{1.0, 0.0}
+	vectorB := []float64{0.0, 1.0}
 	similarity := CosineSimilarity(vectorA, vectorB)
 	if math.Abs(similarity) > 1e-6 {
 		t.Errorf("orthogonal vectors should have similarity 0.0, got %f", similarity)
@@ -23,8 +23,8 @@ func TestCosineSimilarityOrthogonal(t *testing.T) {
 }
 
 func TestCosineSimilarityOpposite(t *testing.T) {
-	vectorA := []float32{1.0, 0.0}
-	vectorB := []float32{-1.0, 0.0}
+	vectorA := []float64{1.0, 0.0}
+	vectorB := []float64{-1.0, 0.0}
 	similarity := CosineSimilarity(vectorA, vectorB)
 	if math.Abs(similarity+1.0) > 1e-6 {
 		t.Errorf("opposite vectors should have similarity -1.0, got %f", similarity)
@@ -32,8 +32,8 @@ func TestCosineSimilarityOpposite(t *testing.T) {
 }
 
 func TestCosineSimilarityMismatchedLength(t *testing.T) {
-	vectorA := []float32{1.0, 2.0}
-	vectorB := []float32{1.0, 2.0, 3.0}
+	vectorA := []float64{1.0, 2.0}
+	vectorB := []float64{1.0, 2.0, 3.0}
 	similarity := CosineSimilarity(vectorA, vectorB)
 	if similarity != 0 {
 		t.Errorf("mismatched length should return 0, got %f", similarity)
@@ -41,15 +41,15 @@ func TestCosineSimilarityMismatchedLength(t *testing.T) {
 }
 
 func TestCosineSimilarityEmpty(t *testing.T) {
-	similarity := CosineSimilarity([]float32{}, []float32{})
+	similarity := CosineSimilarity([]float64{}, []float64{})
 	if similarity != 0 {
 		t.Errorf("empty vectors should return 0, got %f", similarity)
 	}
 }
 
 func TestCosineSimilarityZeroVector(t *testing.T) {
-	vectorA := []float32{0.0, 0.0}
-	vectorB := []float32{1.0, 2.0}
+	vectorA := []float64{0.0, 0.0}
+	vectorB := []float64{1.0, 2.0}
 	similarity := CosineSimilarity(vectorA, vectorB)
 	if similarity != 0 {
 		t.Errorf("zero vector should return 0, got %f", similarity)
